requirements.txt
feedparser
anthropic
python-telegram-bot
apscheduler
psycopg2-binary
